+++ libs/hitables/Sphere.cpp
@@ -1,19 +1,19 @@
 #include "Sphere.h"
-bool Sphere::hit(const Ray & r, float t_min, float t_max, hit_record & rec) const {
+bool Sphere::hit(const Ray & r, double t_min, double t_max, hit_record & rec) const {
 
     //bhaskara to see if the ray collides with the sphere
 
     Vector3 oc = r.get_origin() - center;
 
-    float a = dot(r.get_direction(), r.get_direction());
-    float b = 2 * dot(oc, r.get_direction());
-    float c = dot(oc,oc) - radius*radius;
+    double a = dot(r.get_direction(), r.get_direction());
+    double b = 2 * dot(oc, r.get_direction());
+    double c = dot(oc,oc) - radius*radius;
 
-    float delta = b*b - 4*a*c;
+    double delta = b*b - 4*a*c;
 
     if (delta >= 0) { //means that the ray colides with the sphere
-        float t1 = (-b - sqrt(delta))/(2.0*a);
-        float t2 = (-b + sqrt(delta))/(2.0*a);
+        double t1 = (-b - sqrt(delta))/(2.0*a);
+        double t2 = (-b + sqrt(delta))/(2.0*a);
 
         //now we try to get the minimum positive of the t's
         //which is closests to the ray origin.
